+++ src/components/theme/Logo/Logo.jsx
@@ -5,8 +5,10 @@
 
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { defineMessages, injectIntl } from 'react-intl';
+import { defineMessages, useIntl } from 'react-intl';
 import { Image } from 'semantic-ui-react';
+import { useSelector } from 'react-redux';
+import { settings } from '~/config';
 
 import LogoImage from '@plone/volto/components/theme/Logo/Logo.svg';
 
@@ -27,15 +29,23 @@ const messages = defineMessages({
  * @param {Object} intl Intl object
  * @returns {string} Markup of the component.
  */
-const Logo = ({ intl }) => (
-  <Link to="/" title={intl.formatMessage(messages.site)}>
-    <Image
-      src={LogoImage}
-      alt={intl.formatMessage(messages.plonesite)}
-      title={intl.formatMessage(messages.plonesite)}
-      height={64}
-    />
-  </Link>
-);
+const Logo = () => {
+  const lang = useSelector(state => state.intl.locale);
+  const intl = useIntl();
 
-export default injectIntl(Logo);
+  return (
+    <Link
+      to={settings.isMultilingual ? `/${lang}` : '/'}
+      title={intl.formatMessage(messages.site)}
+    >
+      <Image
+        src={LogoImage}
+        alt={intl.formatMessage(messages.plonesite)}
+        title={intl.formatMessage(messages.plonesite)}
+        height={64}
+      />
+    </Link>
+  );
+};
+
+export default Logo;