+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 
 ## 4.1.2 (2020-03-30)
+- /contents (folder content view) is sortable @ksuess
 
 ### Changes