+++ CHANGELOG.md
@@ -4,12 +4,11 @@
 
 ### Added
 
-- /contents view is sortable @ksuess
+- /contents (folder content view) is sortable @ksuess
 
 ### Changes
 
 ## 4.1.2 (2020-03-30)
-- /contents (folder content view) is sortable @ksuess
 
 ### Changes